[+] Added function to convert DapMath to double.

+++ core/src/dap_math_convert.c
@@ -527,6 +527,6 @@ uint128_t dap_uint128_scan_decimal(const char *a_str_decimal)
 double dap_uint256_decimal_to_double(uint256_t a_decimal){
     uint64_t l_test = dap_uint256_to_uint64(a_decimal);
     uint64_t base = 1000000000000000000;
-    double l_result = (double)(l_test * base);
+    double l_result = (double)((double)l_test * (double)base);
     return l_result;
 }
